Improved category and tags display.
Categories are now defined in config.toml. They are rendered without modification in all views. Tags are defined dynamically. They are rendered via 'humanize | lower' to homogenize their appearance. Categories and tags are hidden from the card view of posts on the main page, but show up in the click-through single page view for each article."